+++ example/app/testapp/urls.py
@@ -3,4 +3,5 @@ from testapp import views
 
 urlpatterns = [
     path("", views.index),
+    path("favicon.ico", views.favicon),
 ]