+++ src/apig_wsgi.py
@@ -13,7 +13,7 @@ DEFAULT_NON_BINARY_CONTENT_TYPE_PREFIXES = (
 
 
 def make_lambda_handler(
-    wsgi_app, binary_support=False, non_binary_content_type_prefixes=None
+    wsgi_app, binary_support=None, non_binary_content_type_prefixes=None
 ):
     """
     Turn a WSGI app callable into a Lambda handler function suitable for
@@ -35,12 +35,27 @@ def make_lambda_handler(
     non_binary_content_type_prefixes = tuple(non_binary_content_type_prefixes)
 
     def handler(event, context):
-        environ = get_environ(event, context, binary_support=binary_support)
-        response = Response(
-            binary_support=binary_support,
-            non_binary_content_type_prefixes=non_binary_content_type_prefixes,
-            multi_value_headers=environ["apig_wsgi.multi_value_headers"],
-        )
+        # Assume version 1 since ALB isn't documented as sending a version.
+        version = event.get("version", "1.0")
+        if version == "1.0":
+            # Binary support deafults 'off' on version 1
+            event_binary_support = binary_support or False
+            environ = get_environ_v1(
+                event, context, binary_support=event_binary_support
+            )
+            response = V1Response(
+                binary_support=event_binary_support,
+                non_binary_content_type_prefixes=non_binary_content_type_prefixes,
+                multi_value_headers=environ["apig_wsgi.multi_value_headers"],
+            )
+        elif version == "2.0":
+            environ = get_environ_v2(event, context, binary_support=binary_support)
+            response = V2Response(
+                binary_support=True,
+                non_binary_content_type_prefixes=non_binary_content_type_prefixes,
+            )
+        else:
+            raise ValueError("Unknown version {!r}".format(event["version"]))
         result = wsgi_app(environ, response.start_response)
         response.consume(result)
         return response.as_apig_response()
@@ -48,20 +63,14 @@ def make_lambda_handler(
     return handler
 
 
-def get_environ(event, context, binary_support):
-    method = event["httpMethod"]
-    body = event.get("body", "") or ""
-    if event.get("isBase64Encoded", False):
-        body = b64decode(body)
-    else:
-        body = body.encode("utf-8")
-
+def get_environ_v1(event, context, binary_support):
+    body = get_body(event)
     environ = {
         "CONTENT_LENGTH": str(len(body)),
         "HTTP": "on",
         "PATH_INFO": event["path"],
         "REMOTE_ADDR": "127.0.0.1",
-        "REQUEST_METHOD": method,
+        "REQUEST_METHOD": event["httpMethod"],
         "SCRIPT_NAME": "",
         "SERVER_PROTOCOL": "HTTP/1.1",
         "SERVER_NAME": "",
@@ -113,30 +122,78 @@ def get_environ(event, context, binary_support):
         # Multi-value headers accumulate with ","
         environ["HTTP_" + key] = ",".join(values)
 
-    # Pass the AWS requestContext to the application
     if "requestContext" in event:
         environ["apig_wsgi.request_context"] = event["requestContext"]
-
-    # Pass the full event to the application as an escape hatch
     environ["apig_wsgi.full_event"] = event
     environ["apig_wsgi.context"] = context
 
     return environ
 
 
-class Response(object):
-    def __init__(
-        self,
-        binary_support,
-        non_binary_content_type_prefixes,
-        multi_value_headers=False,
-    ):
+def get_environ_v2(event, context, binary_support):
+    body = get_body(event)
+    headers = event["headers"]
+    http = event["requestContext"]["http"]
+    environ = {
+        "CONTENT_LENGTH": str(len(body)),
+        "HTTP": "on",
+        "HTTP_COOKIE": ";".join(event.get("cookies", ())),
+        "PATH_INFO": http["path"],
+        "QUERY_STRING": event["rawQueryString"],
+        "REMOTE_ADDR": http["sourceIp"],
+        "REQUEST_METHOD": http["method"],
+        "SCRIPT_NAME": "",
+        "SERVER_NAME": "",
+        "SERVER_PORT": "",
+        "SERVER_PROTOCOL": http["protocol"],
+        "wsgi.errors": sys.stderr,
+        "wsgi.input": BytesIO(body),
+        "wsgi.multiprocess": False,
+        "wsgi.multithread": False,
+        "wsgi.run_once": False,
+        "wsgi.version": (1, 0),
+        "wsgi.url_scheme": "http",
+        # For backwards compatibility with apps upgrading from v1
+        "apig_wsgi.multi_value_headers": False,
+    }
+
+    for key, raw_value in headers.items():
+        key = key.upper().replace("-", "_")
+        if key == "CONTENT_TYPE":
+            environ["CONTENT_TYPE"] = raw_value.split(",")[-1]
+        elif key == "HOST":
+            environ["SERVER_NAME"] = raw_value.split(",")[-1]
+        elif key == "X_FORWARDED_PROTO":
+            environ["wsgi.url_scheme"] = raw_value.split(",")[-1]
+        elif key == "X_FORWARDED_PORT":
+            environ["SERVER_PORT"] = raw_value.split(",")[-1]
+        elif key == "COOKIE":
+            environ["HTTP_COOKIE"] += ";" + raw_value
+            continue
+
+        environ["HTTP_" + key] = raw_value
+
+    environ["apig_wsgi.request_context"] = event["requestContext"]
+    environ["apig_wsgi.full_event"] = event
+    environ["apig_wsgi.context"] = context
+
+    return environ
+
+
+def get_body(event):
+    body = event.get("body", "") or ""
+    if event.get("isBase64Encoded", False):
+        return b64decode(body)
+    return body.encode("utf-8")
+
+
+class BaseResponse:
+    def __init__(self, *, binary_support, non_binary_content_type_prefixes):
         self.status_code = 500
         self.headers = []
         self.body = BytesIO()
         self.binary_support = binary_support
         self.non_binary_content_type_prefixes = non_binary_content_type_prefixes
-        self.multi_value_headers = multi_value_headers
 
     def start_response(self, status, response_headers, exc_info=None):
         if exc_info is not None:
@@ -154,24 +211,6 @@ class Response(object):
             close = getattr(result, "close", None)
             if close:
                 close()
-
-    def as_apig_response(self):
-        response = {"statusCode": self.status_code}
-        # Return multiValueHeaders as header if support is required
-        if self.multi_value_headers:
-            headers = defaultdict(list)
-            [headers[k].append(v) for k, v in self.headers]
-            response["multiValueHeaders"] = dict(headers)
-        else:
-            response["headers"] = dict(self.headers)
-
-        if self._should_send_binary():
-            response["isBase64Encoded"] = True
-            response["body"] = b64encode(self.body.getvalue()).decode("utf-8")
-        else:
-            response["body"] = self.body.getvalue().decode("utf-8")
-
-        return response
 
     def _should_send_binary(self):
         """
@@ -200,3 +239,56 @@ class Response(object):
         if len(matching_headers):
             return matching_headers[-1]
         return None
+
+
+class V1Response(BaseResponse):
+    def __init__(self, *, multi_value_headers, **kwargs):
+        super().__init__(**kwargs)
+        self.multi_value_headers = multi_value_headers
+
+    def as_apig_response(self):
+        response = {"statusCode": self.status_code}
+        # Return multiValueHeaders as header if support is required
+        if self.multi_value_headers:
+            headers = defaultdict(list)
+            [headers[k].append(v) for k, v in self.headers]
+            response["multiValueHeaders"] = dict(headers)
+        else:
+            response["headers"] = dict(self.headers)
+
+        if self._should_send_binary():
+            response["isBase64Encoded"] = True
+            response["body"] = b64encode(self.body.getvalue()).decode("utf-8")
+        else:
+            response["isBase64Encoded"] = False
+            response["body"] = self.body.getvalue().decode("utf-8")
+
+        return response
+
+
+class V2Response(BaseResponse):
+    def as_apig_response(self):
+        response = {
+            "statusCode": self.status_code,
+        }
+
+        headers = {}
+        cookies = []
+        for key, value in self.headers:
+            key_lower = key.lower()
+            if key_lower == "set-cookie":
+                cookies.append(value)
+            else:
+                headers[key_lower] = value
+
+        response["cookies"] = cookies
+        response["headers"] = headers
+
+        if self._should_send_binary():
+            response["isBase64Encoded"] = True
+            response["body"] = b64encode(self.body.getvalue()).decode("utf-8")
+        else:
+            response["isBase64Encoded"] = False
+            response["body"] = self.body.getvalue().decode("utf-8")
+
+        return response